PythonTools/requirements.txt
openai
requests
beautifulsoup4
pandas
python-dotenv
flask
numpy
scikit-learn
google-generativeai
